+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v0.5.3
+- Display results in "Classic', "VirtualText", "Floating Windows", and "Terminal" mode
+- Checkhealth pings binary to check compatibility
+- Official Mac (incl. M1) support
+- Fix for the AUR packaged version
+
+
 ## v0.5.2
 - Ada and Scala interpreters
 - Operator mode, courtesy of @DarwinSenior